axi-hdmi-rx.c: Support 24-bit bus between chip and logic
To use a 24-bit bus, set the "bus-width" property of the endpoint
to <24>. This makes the "CSC_BYPASS" bit toggle between "16" and "24"
bit modes for logic. When using 24-bit mode, the YUV-RGB conversion
is not used in logic, the data is passed in RGB mode from chip to
logic directly.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>