iio: ad9528: Automatically compute N divider A and B counter settings
For the VCO calibration to succeed the calibration divider A and B counter
settings must fulfill the following equation

	B * 4 + A = M1 * N2

Currently the A and B counter settings need to be manually specified. This
is error-prone and unnecessary.

Simply compute them in the driver based on the M1 and N2 settings. This way
they will always be correct and there is no room for accidental
misconfiguration.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>