iio: ad9528: Fix updating clock output rate
When updating the register with the new output clock divider the LSB of
previous value is not properly cleared due to how the mask is computed.

E.g. going from a divider value of 0x1 to 0x2 will result in a value of 0x3
to be written to the register, resulting in the wrong divider settings.

Fix this by explicitly defining the mask rather then trying to use some
macro magic. Do this for all register fields that are updated for
consistency and to avoid future problems.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>