usb: dwc3: gadget: Don't give back URB when ep_queue() fails
When the __dwc3_gadget_kick_transfer() operation fails in
__dwc3_gadget_ep_queue() the function will return an error. This means that
from a caller perspective the URB was never queued. But still
__dwc3_gadget_kick_transfer() will also give back the URB as if it was
completed.

This typically results in a double free on the caller side since it will
free the URB when usb_ep_queue fails and does not expect the complete
callback to be invoked either.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>