drivers: jesd204: add initial skeleton framework
This change adds the initial skeleton for the JESD204 subsystem.
Functionality for this framework will be added in the next patches, this
change adds the basic hooks for init & uninit of the framework.

Inspiration for this code has been taken from the clock framework and some
parts of the IIO subsystem.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>