iio: adc: ad7124: Align with upstream version
The ad7124 driver deviates from the upstream version which was modified
according to the feedback received during review.

The most important difference is regarding the properties that were
added in the devicetree. The gain, and odr-hz properties did not belong
in the DT, while bipolar and diff-channels were defined as generic
properties.

Another change is that the scale can be set from the user space.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>